api/requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
sseclient-py
requests
pydantic